# Faster string checking