--- a/index.js
+++ b/index.js
@@ -8,17 +8,22 @@
 'use strict';
 
 module.exports = function isNumber(num) {
-  var type = typeof num;
-
-  if (type === 'string') {
-    // both a non-empty whitespace string and an empty string will be coerced to 0
-    // with the logic below. By stripping whitespace we can guarantee a correct result.
-    if (num.trim() === '') {
-      return false;
-    }
-  } else if (type !== 'number') {
+  var number = +num;
+  
+  if ((number - number) !== 0) {
+    // Discard Infinity and NaN
     return false;
   }
-
-  return (num - num + 1) === 1;
+  
+  if (number === num) {
+    return true;
+  } else if (typeof num === 'string') {
+    // String parsed, both a non-empty whitespace string and an empty string
+    // will have been coerced to 0. If 0 trim the string and see if its empty.
+    if (number === 0 && num.trim() === '') {
+      return false;
+    }
+    return true;
+  } 
+  return false;
 };