@jonschlinkert 2018-03-31
published and released as 6.0. thanks again!